requirements.txt
asgiref==3.11.0
bcrypt==4.2.1
DateTime==6.0
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
django-rest-framework==0.1.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
mysql-connector==2.2.9
mysqlclient==2.2.7
packaging==26.1
pillow==12.2.0
PyJWT==2.12.0
pytz==2026.1.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
drf-spectacular==0.28.0
zope.interface==8.2
